Only add/remove points based on facemesh predictions when receiving them
...This doesn't really do what I want it to do. The problem is the latency of the results coming from the worker,
and naming a variable that says it's fresh doesn't exactly make it fresh. It might help, though?
It might help to store the optical flow movement over the period between sending the image data and receiving the face data, and applying that to the face data (the scaledMesh and annotations)