Only apply naive latency compensation if previous prediction was confident
This doesn't cover all cases, and is still not necessarily good overall when the latency is bad.
I'm using "if previous prediction was confident" as a stand-in for "if it's tracking your head".
The latency compensation currently can't help with initially tracking your head, which is the most important thing, because if it's already tracking, the dots can handle it, we don't need the face prediction (until it loses tracking again).