Change checkbox structure for JS Paint
In JS Paint, which uses 98.css (or styles copied from it), the checkboxes are invisible when implemented with a label surrounding the input element. This fixes that.

This introduces an id, which is annoying, but the API is already like a singleton, assuming a single instance of the UI. I don't think it would make sense to use the UI multiple times for anything other than perhaps testing (but even then it shouldn't be necessary.)