Update references to the core library in website
I was going to make the website a workspace, but I guess
I'm going to have to add some sort of copying step if I want to do that,
while still using GitHub Pages.
If I configure GitHub Pages to deploy a subfolder, I won't be able to
use "../core/*" to point to the core, which would be a sibling folder.
Symlinks might work.