Replace (most of) an outdated comment
Since I've set up bidirectional communication for the CLI output, it's
not like I couldn't have --help print help info from the
already-running app if I wanted, and I've already got --version printing
version info from both processes.
It no longer simply "quits" when the app is already running.

Also squirrel events and --help and --version can be considered the
same "priority" given that they will never normally be passed together.
Other than the fact that squirrel events are part of (un)installation,
and so if there's a chance of errors in any given code, it's better for
it to go first. But that's already commented upon.