Override `eval`+`Function` only for clmtrackr.js
Instead of globally overriding `eval` and the `Function` constructor
with safe versions that work only with `clmtrackr.js`,
export the safe versions to a namespace, and explicitly import them
in `clmtrackr.js`.

This way, if you want a strict CSP, tracky-mouse can still work with
that, but if you want to use `eval` or `new Function` alongside
tracky-mouse, you can do that too, without modifying
`TrackyMouse.loadDependencies` or loading the scripts yourself.
It is simply as if tracky-mouse doesn't use eval, overall.