Restore strict mode in clmtrackr.js
Strict mode affects semantics of `this` and other things in subtle ways,
and may cause bugs that are hard to track down if the whole library was
built and tested using strict mode.

There's not too many direct usage sites of `eval` anyways, so it's not
littering the file with patches, and this actually un-patches the
removal of 'use strict', so it's basically patch-neutral.